cpu/stm32_common: unify stmclk for stm32f[2|4|7] and adapt to use clk_conf
